agente_gerente_operaciones/requirements.txt
openai-agents
openpyxl
pandas
pyyaml
twilio
fastapi
httpx
pytest
